% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{filter_index}
\alias{filter_index}
\title{Filter sPlotOpen Data by positional index}
\usage{
filter_index(data, x, join = FALSE)
}
\arguments{
\item{data}{sPlotOpen data, a named list containing \code{DT} (species composition
data in long format) and \code{header} (plot-level information).}

\item{x}{A vector of positional indices corresponding to the ordered number of plots in sPlotOpen to subset}

\item{join}{If \code{TRUE} (not the default), joins the filtered \code{DT} and \code{header}
tables.}
}
\value{
sPlotOpen data sliced to include only plots at position \code{x}.
If \code{join = F}, a list containing the sliced \code{DT}, \code{header}, and (if applicable) \code{CWM_CWV} tables. If
\code{join = T}, a single data from containing the joined tables.
}
\description{
\code{filter_index()} creates a subset of sPlotOpen data, retaining
plots based on a list of positional indices you specify.
}
\examples{
data(greece)

# create a vector of positional indices used to slice sPlotOpen
mylist <- c(sample(1:nrow(greece$header), 5, replace = FALSE))

# filter dataset to only include points within the polygon
filter_index(greece, mylist, join = FALSE)
}
