% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{filter_polygon}
\alias{filter_polygon}
\title{Filter sPlotOpen Data by Spatial Polygon}
\usage{
filter_polygon(data, x, join = FALSE)
}
\arguments{
\item{data}{sPlotOpen data, a named list containing \code{DT} (species composition
data in long format) and \code{header} (plot-level information).}

\item{x}{A simple features object, path to a shapefile, or matrix with points
(longitude, latitude) in rows.}

\item{join}{If \code{TRUE} (not the default), joins the filtered \code{DT} and \code{header}
tables.}
}
\value{
sPlotOpen data filtered to include only plots that intersect with \code{x}.
If \code{join = F}, a list containing the filtered \code{DT} and \code{header} tables. If
\code{join = T}, a single data from containing the joined tables.
}
\description{
\code{filter_polygon()} creates a geographic subset of sPlotOpen data, retaining
plots within the region (spatial polygon) you specify.
}
\examples{
data(greece)

# create matrix of coordinates defining a polygon
points <- matrix(c(24.4, 41.5, 24.8, 41.5, 24.8, 41.3, 24.4, 41.3, 24.4, 41.5),
                 ncol = 2, byrow = TRUE)

# filter dataset to only include points within the polygon
filter_polygon(greece, points, join = FALSE)
}
