% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{filter_species}
\alias{filter_species}
\title{Filter sPlotOpen Data by Species}
\usage{
filter_species(data, spp_list, join = FALSE)
}
\arguments{
\item{data}{sPlotOpen data, a named list containing \code{DT} (species composition
data in long format) and \code{header} (plot-level information).}

\item{spp_list}{A vector of species names.}

\item{join}{Whether to join the filtered \code{DT} and \code{header} tables.}
}
\value{
sPlotOpen data filtered to include only plots that contain at least
one species in \code{spp_list}. If \code{join = F}, a list containing the filtered
\code{DT} and \code{header} tables. If \code{join = T}, a single data from containing the
joined tables.
}
\description{
\code{filter_species()} subsets the sPlotOpen data, retaining all vegetation plots
that contain at least one species in a list you provide.
}
\examples{
data(greece)

spp <- c("Silene atropurpurea", "Corylus avellana")
greece_filtered <- filter_species(greece, spp, join = FALSE)
}
