% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greece}
\alias{greece}
\title{Greece forests composition data}
\format{
A data frame with 385 rows and 3 columns:
\itemize{
\item PlotObservationID
\item Species
\item Abundance
}
}
\usage{
data(greece)
}
\description{
Community composition data from the "Greece forests" dataset
}
