% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{map_plots}
\alias{map_plots}
\title{Map sPlotOpen Plots}
\usage{
map_plots(data, type = "grid", grid_size = 300, map_extent = "world")
}
\arguments{
\item{data}{Either a list of sPlotOpen tables that includes the \code{header}
table, or the \code{header} table on its own as a data.frame.}

\item{type}{If \code{grid} (the default), plots the number of vegetation plots by
grid cell. If \code{points}, plots the locations of individual vegetation plots.}

\item{grid_size}{The approximate spacing between grid cells (in km) if using
\code{type = grid}.}

\item{map_extent}{The extent of the retured map. Can be \code{world} or \code{aoi}, if the map should be zoomed to the area of interest}
}
\value{
A world map showing the locations of plots.
}
\description{
\code{map_plots()} plots the locations of sPlotOpen Vegetation plots on a world
map.
}
\examples{
\dontrun{
data(greece)
map_plots(greece, grid_size=100, map_extent="aoi")
}
}
