% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{read_sPlot}
\alias{read_sPlot}
\title{Load sPlotOpen data into R}
\usage{
read_sPlot(
  dir = "~/sPlotOpen/data",
  version = "2.0",
  tables = c("header", "DT", "CWM_CWV")
)
}
\arguments{
\item{dir}{Directory where sPlotOpen tables are stored.}

\item{version}{The version of the dataset to be used. Either '1.0' or '2.0' (default).}

\item{tables}{A character vector. Names of tables to load. Options are
(default is to load all):
\itemize{
\item \code{"header"}: plot-level information.
\item \code{"DT"}: data on species composition of each plot in long format.
\item \code{"CWM_CWV"}: community-weighted means and variances for 18 traits.
}}
}
\value{
A named list containing sPlotOpen data tables as tibbles.
}
\description{
\code{read_sPlot()} searches for sPlotOpen data tables in the directory you
specify and, if present, loads them into R.
}
\examples{
\dontrun{
# Load all sPlotOpen tables
db <- read_sPlot(dir = "~/sPlotOpen/data/2.0", tables = c("DT", "header", "CWM_CWV"))
}
}
