% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{site_by_species}
\alias{site_by_species}
\title{Create a Site-by-Species Matrix}
\usage{
site_by_species(data, sparse = FALSE, pres_abs = FALSE)
}
\arguments{
\item{data}{sPlotOpen species composition data in long format.}

\item{sparse}{Whether to return a sparse matrix (default is \code{FALSE}).}

\item{pres_abs}{Whether to convert species relative cover data to
presence-absence (default is \code{FALSE}).}
}
\value{
A matrix (sites as rows, species as columns) of species relative
cover or presence-absence.
}
\description{
Create a Site-by-Species Matrix
}
\examples{
data(greece)

comp <- greece$DT
m <- site_by_species(comp)
}
