% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{site_species}
\alias{site_species}
\title{Create site-by-species matrix}
\usage{
site_species(data, sparse = TRUE)
}
\arguments{
\item{data}{A data frame containing site names in the first column, species names in the second column, and, optionally, abundance data in the third column.}

\item{sparse}{Whether to return a sparse matrix.}
}
\value{
If \code{sparse = TRUE} a sparse matrix, otherwise a matrix.
}
\description{
Create site-by-species matrix
}
