% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_sPlot}
\alias{get_sPlot}
\title{Download sPlotOpen data and load into R}
\usage{
get_sPlot(dir = NULL, load = TRUE)
}
\arguments{
\item{dir}{character; Directory where sPlotOpen data will be saved after download. If NULL, data will not be saved on disk and only loaded into the R environment.}

\item{load}{logical; Should the data be loaded immediately into R?}
}
\value{
If \code{load = TRUE}, returns a list containing sPlotOpen data tables.
}
\description{
Download sPlotOpen data and load into R
}
