% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greece}
\alias{greece}
\title{Greece forests composition data}
\format{
A list containing:
\itemize{
\item \code{DT}, a data frame with species composition data in long format (385 rows x 6 columns)
\item \code{header}, a data frame with plot-level information (17 rows x 47 columns)
}
}
\usage{
data(greece)
}
\description{
Community composition data from the "Greece forests" dataset
}
