% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{read_sPlot}
\alias{read_sPlot}
\title{Load sPlotOpen data into R}
\usage{
read_sPlot(dir = "~/sPlotOpen/data")
}
\arguments{
\item{dir}{Directory where sPlotOpen tables are stored.}
}
\value{
List containing sPlotOpen data tables.
}
\description{
Load sPlotOpen data into R
}
