% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{site_by_species}
\alias{site_by_species}
\title{Create a Site-by-Species Matrix}
\usage{
site_by_species(data, sparse = FALSE, pres_abs = FALSE)
}
\arguments{
\item{data}{Either a list of sPlotOpen tables that includes the \code{DT} table,
or the \code{DT} table on its own as a data.frame.}

\item{sparse}{If \code{TRUE} (not the default), returns a sparse matrix.}

\item{pres_abs}{If \code{TRUE} (not the default), converts species relative cover
data to presence-absence.}
}
\value{
A matrix (sites as rows, species as columns) of species relative
cover or presence-absence.
}
\description{
\code{site_by_species()} converts sPlotOpen species composition data in long
format (\code{DT} table) to a site-by-species abundance or presence-absence
matrix.
}
\examples{
data(greece)

comp <- greece$DT
m <- site_by_species(comp)
}
