% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{site_species}
\alias{site_species}
\title{Create a Site-by-Species Matrix}
\usage{
site_species(data, sparse = FALSE)
}
\arguments{
\item{data}{A data frame containing site names in the first column, species names in the second column, and, optionally, abundance data in the third column.}

\item{sparse}{Whether to return a sparse matrix.}
}
\value{
If \code{sparse = TRUE} a sparse matrix, otherwise a matrix.
}
\description{
Create a Site-by-Species Matrix
}
\examples{
data(greece)

comp <- greece$DT[,c('PlotObservationID', 'Species', 'Relative_cover')]
m <- site_species(comp)
}
